Create README.md
## Results

After running the tool, comprehensive information about CORS policies and any associated vulnerabilities will be examined. The results are displayed as text messages.

## Contribution

You can contribute to the improvement and development of this tool. To submit suggestions or report issues, you can create a new Issue or send a Pull Request.

## License

This project is licensed under the MIT License. For more information, please read the LICENSE file.